bare_metal - ATmega328P - Arduino - ADC

+++ embedded/bare_metal/ATmega328P (Arduino)/1. blink/src/main.c
@@ -19,7 +19,7 @@
 #include "my_millis.h"
 
 uint32_t last_blink_time = 0;
-const int BLINK_DURATION = 1000;
+const uint16_t BLINK_DURATION = 1000;
 
 int main()
 {
